+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Version 0.7.x-dev
 - fixed default value for option wdayMaxLen
 - renamed package to nexendrie/event-calendar
+- BC break: removed BasicCalendar
 
 Version 0.6.1
 - initialized AbstractCalender::$onDateChange